Move services and controllers into subfolders for contacts app
This uses an extra level of folders to separate out controllers
and services from the rest of the files within the sub-app.
