+++ src/apps/contacts/controllers/index.js
@@ -1,0 +1,11 @@
+const details = require('./controllers/details.controller')
+const interactions = require('./controllers/interactions.controller')
+const edit = require('./controllers/edit.controller')
+const archive = require('./controllers/archive.controller')
+
+module.exports = {
+  details,
+  interactions,
+  edit,
+  archive,
+}
