+++ src/apps/contacts/services/index.js
@@ -1,0 +1,9 @@
+const data = require('./services/data.service')
+const form = require('./services/form.service')
+const formatting = require('./services/formatting.service')
+
+module.exports = {
+  data,
+  form,
+  formatting,
+}
