+++ test/unit/controllers/contact-archive.controller.test.js
@@ -14,8 +14,8 @@ describe('Contact controller, archive', function () {
   beforeEach(function () {
     contactRepositoryArchiveContactStub = sinon.stub().resolves(null)
     contactRepositoryUnArchiveContactStub = sinon.stub().resolves(null)
-    contactArchiveController = proxyquire('~/src/apps/contacts/contact-archive.controller', {
-      './contacts.repo': {
+    contactArchiveController = proxyquire('~/src/apps/contacts/controllers/archive.controller', {
+      '../contacts.repo': {
         archiveContact: contactRepositoryArchiveContactStub,
         unarchiveContact: contactRepositoryUnArchiveContactStub,
       },
@@ -110,8 +110,8 @@ describe('Contact controller, archive', function () {
   it('should handle errors when you archive a contact', function (done) {
     const error = Error('error')
     contactRepositoryArchiveContactStub = sinon.stub().rejects(error)
-    contactArchiveController = proxyquire('~/src/apps/contacts/contact-archive.controller', {
-      './contacts.repo': {
+    contactArchiveController = proxyquire('~/src/apps/contacts/controllers/archive.controller', {
+      '../contacts.repo': {
         archiveContact: contactRepositoryArchiveContactStub,
         unarchiveContact: contactRepositoryUnArchiveContactStub,
       },
@@ -136,8 +136,8 @@ describe('Contact controller, archive', function () {
   it('should handle errors when you unarchive a contact', function (done) {
     const error = Error('error')
     contactRepositoryUnArchiveContactStub = sinon.stub().rejects(error)
-    contactArchiveController = proxyquire('~/src/apps/contacts/contact-archive.controller', {
-      './contacts.repo': {
+    contactArchiveController = proxyquire('~/src/apps/contacts/controllers/archive.controller', {
+      '../contacts.repo': {
         archiveContact: contactRepositoryArchiveContactStub,
         unarchiveContact: contactRepositoryUnArchiveContactStub,
       },
