+++ test/unit/controllers/contact-edit.controller.test.js
@@ -24,18 +24,18 @@ describe('Contact controller, edit', function () {
     getContactAsFormDataStub = sinon.stub().returns({ id: '1234', name: 'Thing' })
     saveContactFormStub = sinon.stub().returns({ id: '1234', first_name: 'Fred', last_name: 'Smith' })
 
-    contactEditController = proxyquire('~/src/apps/contacts/contact-edit.controller', {
-      './contact-form.service': {
+    contactEditController = proxyquire('~/src/apps/contacts/controllers/edit.controller', {
+      '../services/form.service': {
         getContactAsFormData: getContactAsFormDataStub,
         saveContactForm: saveContactFormStub,
       },
-      '../../repos/metadata.repo': {
+      '../../../repos/metadata.repo': {
         countryOptions: [{
           id: '986',
           name: 'United Kingdom',
         }],
       },
-      '../../repos/company.repo': {
+      '../../../repos/company.repo': {
         getDitCompany: getDitCompanyStub,
       },
     })
@@ -344,18 +344,18 @@ describe('Contact controller, edit', function () {
         error: { name: ['test'] },
       })
 
-      contactEditController = proxyquire('~/src/apps/contacts/contact-edit.controller', {
-        './contact-form.service': {
+      contactEditController = proxyquire('~/src/apps/contacts/controllers/edit.controller', {
+        '../services/form.service': {
           getContactAsFormData: getContactAsFormDataStub,
           saveContactForm: saveContactFormStub,
         },
-        '../../repos/metadata.repo': {
+        '../../../repos/metadata.repo': {
           countryOptions: [{
             id: '986',
             name: 'United Kingdom',
           }],
         },
-        '../../repos/company.repo': {
+        '../../../repos/company.repo': {
           getDitCompany: getDitCompanyStub,
         },
       })
@@ -370,18 +370,18 @@ describe('Contact controller, edit', function () {
     it('should show errors when the save fails for a non-validation related reason', function (done) {
       saveContactFormStub = sinon.stub().rejects(Error('some error'))
 
-      contactEditController = proxyquire('~/src/apps/contacts/contact-edit.controller', {
-        './contact-form.service': {
+      contactEditController = proxyquire('~/src/apps/contacts/controllers/edit.controller', {
+        '../services/form.service': {
           getContactAsFormData: getContactAsFormDataStub,
           saveContactForm: saveContactFormStub,
         },
-        '../../repos/metadata.repo': {
+        '../../../repos/metadata.repo': {
           countryOptions: [{
             id: '986',
             name: 'United Kingdom',
           }],
         },
-        '../../repos/company.repo': {
+        '../../../repos/company.repo': {
           getDitCompany: getDitCompanyStub,
         },
       })
