+++ test/unit/services/contact-form.service.test.js
@@ -8,8 +8,8 @@ describe('contact form service', function () {
   beforeEach(function () {
     throwError = false
 
-    contactFormService = proxyquire('~/src/apps/contacts/contact-form.service', {
-      './contacts.repo': {
+    contactFormService = proxyquire('~/src/apps/contacts/services/form.service', {
+      '../contacts.repo': {
         savedContactForm: null,
         saveContact: function (token, contactForm) {
           return new Promise((resolve, reject) => {
