+++ test/unit/services/contact-formatting.service.test.js
@@ -1,4 +1,4 @@
-const contactFormattingService = require('~/src/apps/contacts/contact-formatting.service')
+const contactFormattingService = require('~/src/apps/contacts/services/formatting.service')
 
 describe('Contact formatting service', function () {
   let contact
